Enable memory info in app info settings
Added config_show_app_info_settings_memory flag to config.xml to make memory information available in app info settings.

overlay: Settings: Show memory usage in app info

The system tracks memory usage for each app, so we might as well show it
to the user so that users can track down which apps are hogging their
memory. This shouldn't be a developer-only feature that was previously
only exposed in Developer Options.

Change-Id: Ib900438f856369286b7621208020c5c8da73765f